jit: Initial commit for LLVM backend
Lets see how hard it is to write a LLVM backend for Jato...

You can enable it with the '-Xllvm' command line option. `EntryTest`
aborts like this:

  [penberg@tux jato]$ ./jato -Xllvm -cp test/functional jvm/EntryTest
  jato: jit/llvm/core.c:325: llvm_bc2ir_insn: Assertion `0' failed.
  Aborted

which is caused by missing bytecode translation for OPC_NEW. That's
hardly surprising as we don't translate any bytecodes. We do, however,
generate code with full LLVM stack that rest of Jato's JIT is able to
call.

Signed-off-by: Pekka Enberg <penberg@kernel.org>