jit/llvm: Use vm_method_arg_stack_count() llvm_function_type()
Use the newly added vm_method_arg_stack_count() helper in
llvm_function_type() and drop the open-coded (and buggy) version.

Please note that this leaves a bug in llvm_function_type() which does
not handle non-static methods nor 64-bit integer or float types
properly. I added an assert() guard that catches the cases so we can fix
them when we actually hit the problem.

Signed-off-by: Pekka Enberg <penberg@kernel.org>